[hannk] Legalize broadcast binary ops
For binary ops that handle broadcasts, only the second input is allowed to be a broadcast. This adds a transforms that legalizes the operand before execution time by swapping the first and second op for those cases. Also adds some handcrafted .tflite files that exercise these cases. (We'd fail with a halide assertion before.)