Update HelloWasm
Recent updates to Chrome/FF/etc require serving pages with certain headers if you want to allow SharedArrayBuffer (and thus, multithreading); update the README and add a helper script to make this documented and easier to manage.

Also add $(CXXFLAGS) to the benchmark targets so they enable C++11 or more (which HalideBuffer.h now requires).