[hannk] Clean up aliasing (v2) (#6364)
* wip

* [hannk] Clean up aliasing (v2)

The code for aliasing tensors was janky. This cleans it up and makes a clear distinction between aliasing done to overlay buffers with crop-and-translate, vs the aliasing done when we reshape tensors. We no longer allow a given tensor to do both of these, and we give preference to Reshape aliasing first.

(Cherry-picked from  #6321)

* Move alias_type into shared ptr