Support vectorization in OpenGLCompute backend (#6348)
* Support vectorization in OpenGLCompute backend

This patch adds support for vector load and store operations.
First, a pass identifies the buffers whose loads and stores are all
dense, aligned, and have the same number of lanes. Such buffers are
declared with a vector base type and accessed accordingly. Loads and
stores that do not satisfy those criteria are implemented as gathers and
scatters from buffers whose base type is scalar.

Resolves #4976.
Partially resolves #1687.

* Move buffer name instead of copy (clang-tidy)