Remove halide_abort_if_false() usage in runtime/metal
This converts all the usage of `halide_abort_if_false()` in runtime/metal into either an explicit runtime check-and-return-error-code (if the check looks plausible), or `halide_debug_assert()` (if the check seems to be stating an invariant that shouldn't be possible in well-structured code). These changes are admittedly subjective, so feedback is especially welcome.

Also, driveby change to sync-common.h to use `halide_debug_assert()` rather than a local equivalent.