Fix broken ASAN code
Various changes and merges ended up with us using multiple ASAN passes, which was pretty crashy (we just didn't notice because it isn't tested well enough on our buildbots, but is elsewhere).

I think we really only want to use the ModuleAddressSanitizerPass (not the non-Module version), which is what Clang does.