Don't remap the function name or the target in the metadata (#6430)
The remapping is only intended to be used for output argument(s), not the function name; if you have an output with the same name as the function, you can get the metadata emitted with incorrect information. (And remapping the target string is just silly.)

This is almost impossible to do currently, but if you construct a Generator just right, you can make it happen.