Sketch of AOT-compiled performance tests
This is a sketch of a clunky-but-effective way to make some (simple) tests AOT-compiled without having to greatly increase number of source files or add lots of complex build rules. Not sure if it's a win or not -- opinions wanted. Basically, this puts a Generator and a 'main' test harness into the same file, separated by an ifdef; the CMake build rules know how to build the file multiple ways and connect them together.

I'm honestly torn as to whether the icky, please-don't-use-this-in-real-code approach here is outweighed by the convenience factor of having self-contained tests that will AOT-compile properly.

(BTW, this approach produces results that will build-and-run just fine under Wasm! Although the one example I've provided in this PR doesn't do so usefully, because it requires GPU support...)

Also note that I haven't provided build rules for Make (only CMake), and I don't intend to do so, because doing this in Make is a big mess of ugly special cases. Someone else can add new features to Make if they like.