[WebGPU] Implement core of WGSL codegen
This implements enough of the WGSL codegen required to generate code
for a 32-bit version of the apps/blur example.

Buffer arguments are emitted as read_write storage buffers.

Non-buffer arguments are collected into a structure and generated as a
single uniform buffer.

Workgroup sizes are currently required to be constant. This can
relaxed in the future by using pipeline-overridable constants.