Move null check from Printer to halide_string_to_string()
The Printer is (currently) usually inlined into every module, so this check is repeated in multiple chunks of code. Since the goal is to avoid crashing when debugging, let's move it to halide_string_to_string() (which will catch all these, and possibly more) and save some code size.

(Further improvements in Printer code size on the way; this change seems worthy of considering separately.)