[NFC-ish] Finish MSAN handling (#6516)
Somehow, initially i missed that there was MSan support,
so it might be good to actually mention that we don't need to run
any MSan passes here, and that we didn't forget to run them.

Secondly, it seems inconsistent not annotate the functions
with `Attribute::SanitizeMemory`, like we do for others.

I suppose it isn't strictly required, since they are used
to actually drive the instrumentation passes, and we don't
run MSan pass, but they are also used to disable some LLVM optimizations,
and that //might// be important. Or not, but then i suppose
there should be a comment about it?

Co-authored-by: Steven Johnson <srj@google.com>
(cherry picked from commit 7eb9949f14f8c164315f81729c7f2e1b32e06aef)