Rename Output to OutputFileType and deprecated Output (#6568)
* Rename Output to OutputFile and deprecated Output

It would be nice in future if we could promote Generator::Input and
Generator::Output to the Halide namespace. For that to happen, we need
to rename the enum Output first. This adds a new name for it, and
deprecates the old name.

* OutputFile -> OutputType

Also fix python bindings and use the right deprecation macro

* OutputType -> OutputFileType

* clang-format

* Update python tutorials and tests