Timer based profiler (#6642)
* Add support for timer interrupt based profiling, which is useful for bring up
on embedded ("bare metal") systems that may not have a full OS with threads.

* Update runtime_api file with new routines.

* Turn locking back on in timer based profiling case as it can be used
in multiprocessor situations. (Both as an option on systems where
threads would be fine and on embedded systems which don't have time
shared threads but cores are dedicated to the Halide threadpool.)

* Add target flag for timer profiling and extend performance_profiler
test to cover timer profiling.