Upgrade the PyStub code to handle synthetic GeneratorParams
PyStub is the code that lets Python code call a (C++) Generator; this upgrades the code to allow passing in of 'synthetic' GeneratorParams (e.g. buffer.type), which it previously didn't handle.

At the moment this doesn't make a big difference (I'm not aware of ~any meaningful amount of code outside of the Python tests that use PyStub), but this will streamline some possible changes in an experimental branch and make future diffs less painful.