Enable `bugprone-implicit-widening-of-multiplication-result` and fix errors
clang-tidy-13 has a new check, `bugprone-implicit-widening-of-multiplication-result`:

"The check diagnoses instances where a result of a multiplication is implicitly widened, and suggests (with fix-it) to either silence the code by making widening explicit, or to perform the multiplication in a wider type, to avoid the widening afterwards."

This seems like a reasonable thing, in theory, so I enabled it and manually fixed everything it complained about (the `-fix` option didn't work well here). Most of the actual changes are probably inconsequential, but checks of this sort seem appropriate for clang-tidy. What do other folks think?