Finish various missing bits of float16/bfloat16 support
Fixes issues #3709, #3967, #4916. I think this is the meaningful remaining bits (outside of Python) that have been left unanswered for a while.

Note the addition of `halide_float16_t` and `halide_bfloat16_t` in HalideRuntime.h as runtime placeholders that are equivalent to the language-only variants in Float16.h.

Note that there are similar issues in the Python bindings, but I'm going to land those in a subsequent PR to minimize confusion.