Make CodegenC stop pretending it can generate plain C code
The public interface seems to claim you can produce C output (in addition to C++), but this is a lie; our output is C++11 (!) and has been for a while. Remove this pretense and clean up a bit.

Note that the only practical difference I see in the C-vs-C++ code was:
- whether to emit functions as extern "C" vs name-mangled; this was (and is) actually handled by the Target::CPlusPlusMangling feature directly.
- the C++ version emits all temporaries with `const` declarations, which is arguably good.

(No, I'm not going to bother renaming the class or file unless someone insists.)