Add a way to prohibit calls to ImageParam/Param::set
The goal here is to allow us to 'freeze' an ImageParam (or Param) so that it is no longer legal to call set_scalar(), set_buffer(), etc on it. In normal JIT usage, it's unlikely you'd ever want to do this, but for Generators, it would be undesirable for Generator code to (say) cast an input to an ImageParam and bind a Buffer to it directly, since the meaning is only well-defined forJIT usage.

This is in the category of "I don't know of anyone misusing this right now, but I want to disallow it before anyone does, because Hyrum's Law exists."

Also, drive-by fixes for const-correctness and memcpy usage in Parameter.h.