Add GeneratorFactoryProvider to generate_filter_main()
This provides hooks to allow overriding the Generator(s) that generate_filter_main() can use; normally it defaults to the global registry of C++ Generators, but this allows for (e.g.) alternate-language-bindings to selectively override this (e.g. to enumerate only Generators that are visible in that language, etc).

(No visible change in behavior from this PR; this is just cherry-picked from work-in-progress elsewhere to simplify review & merge)