AOT `_metadata()` getters should be multitarget-aware
Multitarget codegen has worked correctly for years, but calling the generated `_metadata()` function (returning `halide_filter_metadata_t*`) has been subtly wrong: the `target` field always contains the baseline target. This PR proposes to smarten the codegen by using the same `can_use_target` logic for the generated metadata getters, so that the `target` field you inspect always matches the one that was used for the actual filter that will be executed.

(This is, admittedly, a rather niche use case, but if nothing else, it is helpful for some debug and test scenarios, as it gives us a simpler way to verify at runtime that the version we are executing is the one we expect.)