[WebGPU] Fix mixed types in buffers
The atomic emulation of narrow types shouldn't use the allocation
type, since the same buffer may be re-used for multiple types. This
means we also sometimes need to perform 32-bit accesses using atomics
as well.

Instead of using the buffer allocation type, we now pre-traverse the
IR to check for accesses that will require emulation, and mark the
corresponding buffer as such.