Remove Generator::value_tracker and friends
This is an internal-to-Generator helper that is used to try to detect certain classes of errors when using GeneratorStubs. To the best of my knowledge, it has ~never found a useful error in all of its existence; combined with the very limited usage of GeneratorStubs, I think this code no longer pays for itself, and should be removed.

(Note that this was never externally visible, thus no deprecation warnings should be necessary.)