[Codegen_LLVM] Define all the things
Long-term plan for LLVM is to get rid of `undef`,
and replace it with zero-initialization,
err, `poison`, because it has nicer semantics.

Everywhere we use `undef` as a placeholder in shuffle
(be it either for a second operand, or undef shuffle mask element),
or as a base 'empty' vector we are about to fully override
via insertelement, we can just switch those to poison nowadays.

The scary part is the `Call::undef` semantics/lowering,
perhaps it will need to be `freeze poison`.