Deinterleave: generalize to be lane mask-based
Currently, it was implemented to only handle "deinterleaving"
(really, removal of not-demanded lanes) of lanes that can be
described by a linear polynominal. While this clearly worked,
there already were some FIXME comments, and it was already
clearly not generic-enough, e.g. it could not recurse into shuffles.

For proper handling of Reinterpret, which either merges adjacent lanes.
or splits lanes, this means that no support is possible at all.