[Codegen_LLVM] Arguments/Loads/Returns are well-defined
https://llvm.org/docs/LangRef.html#well-defined-values:
```
Given a program execution, a value is well defined if the value
does not have an undef bit and is not poison in the execution.
An aggregate value or vector is well defined if its elements
are well defined. The padding of an aggregate isn’t considered,
since it isn’t visible without storing it into memory and
loading it with a different type.
```

Same story as with https://github.com/halide/Halide/pull/6897,
this matches //my// understanding of Halide's semantics,
but this may very well not be what the reality is.

To be noted, even //if// this is correct, i do believe
this has a very high chance to cause issues.
I'm also not sure how this plays with MSan.

Notably, while annotating function arguments/return is trivial,
for load's this only annotates the `load` instructions that we
actually create, i.e. it misses loads in the IR that we precompiled
and then link into our module.