Revise PyGen API to have implicit GeneratorContext (#6935)
* Revise PyGen API to have implicit GeneratorContext

This proposes to revise the Python Generator API to have GeneratorContext *never* be an explicit argument; rather, to have an *implicit* current GeneratorContext that is active. This reduces boilerplate when composing code (e.g. via Generator.call). To temporarily push a new GeneratorContext on the active stack, `with hl.set_context(ctx):` is the appropriate thing to do.

* Update _generator_helpers.py

* Revise to avoid need for set_context()

* clang-tidy