get_test_tmp_dir() should return system tmp, not current dir
If TEST_TMPDIR isn't defined, we return the current directory, which is pretty much *never* the right thing. We should use /tmp or the equivalent.

(Note that on Windows there is the potential degenerate case of GetTempPath() failing, usually under either weird test setups or weird enterprise-configuration setups in which various temp dirs of users are artificially locked down; these are both corner cases that we should deal with as they become problems.)