Prohibit C99 VLA usage in runtime code
AFAICT we aren't doing this in Halide at present, but some experimental code in Google runtime was doing so; this caused some issues with some experimental Clang patches, but also was never really intended to be used in the first place. Adding the flag here to be sure no unintended use creeps back in.

While I was there, took the time to ensure that the flags for runtime are unified across CMake and Make.