Add a terminate_handler to try to report unhandled exceptions
This uses an ugly trick: since a terminate_handler doesn't see the exception that was unhandled, we use a try/(re)throw/catch inside the terminate_handler to sniff at it. This is not standard C++, but apparently GCC and Clang (at least) support it, and it dramatically improves the quality of error reporting for our tests when Halide is built with exceptions enabled, so let's see how it works in practice.